AI_Branch_Manager/requirements.txt
Flask
openai
face_recognition
opencv-python
pytesseract
Pillow